requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=9.0.0

# Model libraries
timm>=0.6.0

# Attack library
torchattacks>=3.4.0

# Utilities
matplotlib>=3.5.0
